requirements.txt
streamlit==1.56.0
pandas==3.0.2
plotly==6.7.0
pillow==12.2.0
python-docx==1.2.0
numpy==1.26.4
pyarrow==23.0.1
